LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/animation-nested-transition-expected.txt

Test

FAIL Animated container size triggers transition assert_equals: expected "rgb(100, 100, 100)" but got "rgb(200, 200, 200)"
